Extract BaseJoinReorderingTest
