+++ presto-main/src/test/java/com/facebook/presto/sql/planner/optimizations/TestTpchJoinReordering.java
@@ -15,18 +15,9 @@
 package com.facebook.presto.sql.planner.optimizations;
 
 import com.facebook.presto.Session;
-import com.facebook.presto.sql.planner.LogicalPlanner;
-import com.facebook.presto.sql.planner.Plan;
-import com.facebook.presto.sql.planner.SimplePlanVisitor;
-import com.facebook.presto.sql.planner.assertions.BasePlanTest;
-import com.facebook.presto.sql.planner.plan.JoinNode;
-import com.facebook.presto.sql.planner.plan.SemiJoinNode;
-import com.facebook.presto.sql.planner.plan.TableScanNode;
-import com.facebook.presto.sql.planner.plan.ValuesNode;
 import com.facebook.presto.testing.LocalQueryRunner;
 import com.facebook.presto.tpch.ColumnNaming;
 import com.facebook.presto.tpch.TpchConnectorFactory;
-import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Resources;
 import io.airlift.tpch.Customer;
@@ -46,18 +37,14 @@ import static com.facebook.presto.sql.planner.plan.JoinNode.Type.INNER;
 import static com.facebook.presto.sql.planner.plan.JoinNode.Type.LEFT;
 import static com.facebook.presto.sql.planner.plan.JoinNode.Type.RIGHT;
 import static com.facebook.presto.tpch.TpchConnectorFactory.TPCH_COLUMN_NAMING_PROPERTY;
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkState;
 import static java.lang.String.format;
-import static java.util.Objects.requireNonNull;
-import static org.testng.Assert.assertEquals;
 
 /**
  * This class tests cost-based optimization rules related to joins. It contains unmodified TPCH queries.
  * This class is using TPCH connector configured in way to mock Hive connector with unpartitioned TPCH tables.
  */
-public class TestJoinReordering
-        extends BasePlanTest
+public class TestTpchJoinReordering
+        extends BaseJoinReorderingTest
 {
     /*
      * CAUTION: The expected plans here are not necessarily optimal yet. Their role is to prevent
@@ -66,7 +53,7 @@ public class TestJoinReordering
      * large amount of data.
      */
 
-    public TestJoinReordering()
+    public TestTpchJoinReordering()
     {
         super(
                 "sf3000.0",
@@ -514,213 +501,6 @@ public class TestJoinReordering
                                         tableScan("customer")),
                                 tableScan("orders")),
                         new Values()));
-    }
-
-    private TableScan tableScan(String tableName)
-    {
-        return new TableScan(format("tpch:%s:%s", tableName, getQueryRunner().getDefaultSession().getSchema().get()));
-    }
-
-    private void assertJoinOrder(String sql, Node expected)
-    {
-        assertEquals(joinOrderString(sql), expected.print());
-    }
-
-    private String joinOrderString(String sql)
-    {
-        Plan plan = plan(sql, LogicalPlanner.Stage.OPTIMIZED_AND_VALIDATED, false);
-
-        JoinOrderPrinter joinOrderPrinter = new JoinOrderPrinter();
-        plan.getRoot().accept(joinOrderPrinter, 0);
-        return joinOrderPrinter.result();
-    }
-
-    private static class JoinOrderPrinter
-            extends SimplePlanVisitor<Integer>
-    {
-        private final StringBuilder stringBuilder = new StringBuilder();
-
-        public String result()
-        {
-            return stringBuilder.toString();
-        }
-
-        @Override
-        public Void visitJoin(JoinNode node, Integer indent)
-        {
-            JoinNode.DistributionType distributionType = node.getDistributionType()
-                    .orElseThrow(() -> new IllegalStateException("Expected distribution type to be present"));
-            if (node.isCrossJoin()) {
-                checkState(node.getType() == INNER && distributionType == REPLICATED, "Expected CROSS join to be INNER REPLICATED");
-                stringBuilder.append(indentString(indent))
-                        .append("cross join:\n");
-            }
-            else {
-                stringBuilder.append(indentString(indent))
-                        .append("join (")
-                        .append(node.getType())
-                        .append(", ")
-                        .append(distributionType)
-                        .append("):\n");
-            }
-
-            return visitPlan(node, indent + 1);
-        }
-
-        @Override
-        public Void visitTableScan(TableScanNode node, Integer indent)
-        {
-            stringBuilder.append(indentString(indent))
-                    .append(node.getTable().getConnectorHandle().toString())
-                    .append("\n");
-            return visitPlan(node, indent + 1);
-        }
-
-        @Override
-        public Void visitSemiJoin(final SemiJoinNode node, Integer indent)
-        {
-            stringBuilder.append(indentString(indent))
-                    .append("semijoin (")
-                    .append(node.getDistributionType().map(SemiJoinNode.DistributionType::toString).orElse("unknown"))
-                    .append("):\n");
-
-            return visitPlan(node, indent + 1);
-        }
-
-        @Override
-        public Void visitValues(ValuesNode node, Integer indent)
-        {
-            stringBuilder.append(indentString(indent))
-                    .append("values\n");
-
-            return null;
-        }
-    }
-
-    private static String indentString(int indent)
-    {
-        return Strings.repeat("    ", indent);
-    }
-
-    private interface Node
-    {
-        void print(StringBuilder stringBuilder, int indent);
-
-        default String print()
-        {
-            StringBuilder stringBuilder = new StringBuilder();
-            print(stringBuilder, 0);
-            return stringBuilder.toString();
-        }
-    }
-
-    private Join crossJoin(Node left, Node right)
-    {
-        return new Join(INNER, REPLICATED, true, left, right);
-    }
-
-    private static class Join
-            implements Node
-    {
-        private final JoinNode.Type type;
-        private final JoinNode.DistributionType distributionType;
-        private final boolean isCrossJoin;
-        private final Node left;
-        private final Node right;
-
-        private Join(JoinNode.Type type, JoinNode.DistributionType distributionType, Node left, Node right)
-        {
-            this(type, distributionType, false, left, right);
-        }
-
-        private Join(JoinNode.Type type, JoinNode.DistributionType distributionType, boolean isCrossJoin, Node left, Node right)
-        {
-            if (isCrossJoin) {
-                checkArgument(distributionType == REPLICATED && type == INNER, "Cross join can only accept INNER REPLICATED join");
-            }
-            this.type = requireNonNull(type, "type is null");
-            this.distributionType = requireNonNull(distributionType, "distributionType is null");
-            this.isCrossJoin = isCrossJoin;
-            this.left = requireNonNull(left, "left is null");
-            this.right = requireNonNull(right, "right is null");
-        }
-
-        @Override
-        public void print(StringBuilder stringBuilder, int indent)
-        {
-            if (isCrossJoin) {
-                stringBuilder.append(indentString(indent))
-                        .append("cross join:\n");
-            }
-            else {
-                stringBuilder.append(indentString(indent))
-                        .append("join (")
-                        .append(type)
-                        .append(", ")
-                        .append(distributionType)
-                        .append("):\n");
-            }
-
-            left.print(stringBuilder, indent + 1);
-            right.print(stringBuilder, indent + 1);
-        }
-    }
-
-    private static class SemiJoin
-            implements Node
-    {
-        private final JoinNode.DistributionType distributionType;
-        private final Node left;
-        private final Node right;
-
-        private SemiJoin(JoinNode.DistributionType distributionType, final Node left, final Node right)
-        {
-            this.distributionType = requireNonNull(distributionType);
-            this.left = requireNonNull(left);
-            this.right = requireNonNull(right);
-        }
-
-        @Override
-        public void print(StringBuilder stringBuilder, int indent)
-        {
-            stringBuilder.append(indentString(indent))
-                    .append("semijoin (")
-                    .append(distributionType.toString())
-                    .append("):\n");
-
-            left.print(stringBuilder, indent + 1);
-            right.print(stringBuilder, indent + 1);
-        }
-    }
-
-    private static class TableScan
-            implements Node
-    {
-        private final String tableName;
-
-        private TableScan(String tableName)
-        {
-            this.tableName = tableName;
-        }
-
-        @Override
-        public void print(StringBuilder stringBuilder, int indent)
-        {
-            stringBuilder.append(indentString(indent))
-                    .append(tableName)
-                    .append("\n");
-        }
-    }
-
-    private static class Values
-            implements Node
-    {
-        @Override
-        public void print(StringBuilder stringBuilder, int indent)
-        {
-            stringBuilder.append(indentString(indent))
-                    .append("values\n");
-        }
     }
 
     private static String tpchQuery(int queryNumber)
